requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
boto3==1.35.80
celery==5.4.0
redis==5.2.1
pytesseract==0.3.13
Pillow==11.0.0
pdf2image==1.17.0
python-dotenv==1.0.1
httpx==0.28.1
